+++ framework/Source/CPTPlotAreaFrame.m
@@ -118,6 +118,7 @@
         plotArea = newPlotArea;
         if ( plotArea ) {
             [self insertSublayer:plotArea atIndex:0];
+            plotArea.graph = self.graph;
         }
         [self setNeedsLayout];
     }
@@ -141,6 +142,15 @@
 -(void)setPlotGroup:(CPTPlotGroup *)newPlotGroup
 {
     self.plotArea.plotGroup = newPlotGroup;
+}
+
+-(void)setGraph:(CPTGraph *)newGraph
+{
+    if ( newGraph != self.graph ) {
+        [super setGraph:newGraph];
+
+        self.plotArea.graph = newGraph;
+    }
 }
 
 /// @endcond